AI-Gateway/labs/master-lab/.archive/old-archive/scripts/requirements.txt
# Master AI Gateway Lab - Python Requirements
# Install with: pip install -r requirements.txt

# Azure SDK - AI Services
azure-ai-inference==1.0.0b9  # Pin to latest beta; stable 1.0.0 not yet published
azure-ai-projects>=1.0.0
azure-identity>=1.15.0
azure-core>=1.29.0
azure-search-documents>=11.4.0

# Azure SDK - Management & Deployment
azure-mgmt-resource>=23.0.0
azure-mgmt-cognitiveservices>=13.0.0
azure-mgmt-apimanagement>=4.0.0
azure-cli-core>=2.50.0

# Azure SDK - Storage & Data
azure-cosmos>=4.5.0
azure-storage-blob>=12.19.0

# OpenAI
openai>=1.12.0

# MCP (Model Context Protocol)
mcp==1.15.0

# Data Science & Visualization
pandas==2.3.3
matplotlib==3.10.6
matplotlib-inline==0.1.7
numpy>=1.24.0

# Image Processing
pillow==11.3.0

# HTTP & Async
requests>=2.31.0
aiohttp>=3.9.0
nest-asyncio==1.6.0

# Redis
redis>=5.0.0

# Utilities
python-dotenv>=1.0.0
